+++ CHANGELOG.md
@@ -607,3 +607,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 
 #### dev ####
 * Fix colored text being offset badly (kiooeht)
+* Fix multi-word keywords not removing underscores (kiooehts)