+++ CHANGELOG.md
@@ -607,4 +607,5 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 
 #### dev ####
 * Fix colored text being offset badly (kiooeht)
-* Fix multi-word keywords not removing underscores (kiooehts)
+* Fix multi-word keywords not removing underscores (kiooeht)
+* Fix adding and removing Watcher cards (kiooeht)