+++ CHANGELOG.md
@@ -605,9 +605,11 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 #### v5.16.0-beta ####
 * Support for Slay the Spire 09-09-2019 BETA (kiooeht)
 
-#### dev ####
+#### v5.16.0-beta.2 ####
 * Fix colored text being offset badly (kiooeht)
 * Fix multi-word keywords not removing underscores (kiooeht)
 * Fix adding and removing Watcher cards (kiooeht)
 * Make Vampires event use standard remove function (kiooeht)
 * Move card preview in SingleCardView to not cover whatmod (kiooeht)
+
+#### dev ####